+++ worker/include/RTC/Router.hpp
@@ -74,7 +74,7 @@ namespace RTC
 		void OnTransportDataProducerSctpMessageReceived(
 		  RTC::Transport* transport,
 		  RTC::DataProducer* dataProducer,
-		  uint8_t ppid,
+		  uint32_t ppid,
 		  const uint8_t* msg,
 		  size_t len) override;
 		void OnTransportNewDataConsumer(
